feat: update package.json and add VitePress configuration
- Set package as private and define module type in package.json
- Upgrade @shikijs/vitepress-twoslash to version 3.9.2
- Add vitepress-plugin-mermaid to dependencies
- Create a new VitePress configuration file with custom settings
- Integrate Tailwind CSS and group icon plugins
- Configure markdown with Twoslash transformer and group icon plugin
- Add custom head elements including scripts for analytics and chat widget
- Define theme configuration with navigation, sidebar, and social links
