+++ package.json
@@ -1,6 +1,8 @@
 {
   "name": "iexec-documentation",
   "version": "0.0.0",
+  "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vitepress dev",
     "build": "vitepress build",
@@ -30,11 +32,12 @@
     "viem": "^2.33.1",
     "vitepress-plugin-google-analytics": "^1.0.2",
     "vitepress-plugin-group-icons": "^1.6.1",
+    "vitepress-plugin-mermaid": "^2.0.17",
     "vue": "^3.5.17"
   },
   "devDependencies": {
     "@iconify/vue": "^5.0.0",
-    "@shikijs/vitepress-twoslash": "^3.8.0",
+    "@shikijs/vitepress-twoslash": "^3.9.2",
     "@types/figlet": "^1.7.0",
     "@types/prop-types": "^15.7.15",
     "@types/react": "^18.3.23",
